don't share query cache across scanners
Cache entries from older searches for different types of things could interfere
with the current search. Stop this by clearing the cache on probe invocation.
We could save the cache across invocations, but we'd need to index the cache by
both the scanning function and the current directory for file finding, and it
doesn't make sense to do so for sets of small objects that change frequently,
like buffers.