add separate caches for repo branches
For VCSes that checkout different branches in the same directory, like git and
hg, it can be nice to have separate caches for different branches so you don't
have to do a full cache refresh everytime you change branches in order to find
the files that are in only one of the branches. I have always got around the
issue by using a long-standing vim process for each branch and using the buffer
search to open windows for the new files in the two branches, but that's
obviously very specific to my personal workflow.

Note that the branch-specific caches are only used if the ProbeFindFileInRepo
command is used. So, ProbeFindFile and ProbeFindFileInRepo might use different
caches, but I don't think it'll be a problem.

In the interests of simplicity I considered having this off by default, but I'm
making it the default since I think it'll will be more intuitive for most
people.