fix opening filepaths that include spaces
You don't want to give an escaped path to fnamemodify: if fnamemodify is
able to shorten it (in this case that would mean the file being opened
is under the working directory), it will convert backslashes to slashes.

Vim automatically shortens the filepath when displaying it in the
statusline, so using fnamemodify was pointless anyway.