use unix-style paths on windows
Using the same path separator on windows and unix lets us share
directory listing caches between cygwin and windows for directories
under $HOME, and makes joining paths work better. It seems like vim can
handle a mix of backslashes and slashes in a path, but it's slow to do
the extra directory lookups.